+++ src/backend/funmooc/auth/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AuthConfig(AppConfig):
+    label = "funmooc.auth"
+    name = "funmooc.auth"